skip TestNoHardwareSource when test -race
moves TestNoHardwareSource to a new file that uses a build tag to skip it when test is called with -race. annoyingly, -race doesn't show up in flag
